// back-end/libs/common/error/rpc/rpc-error.enum.ts
export enum RpcErrorType {
    INVALID_INPUT = 'INVALID_INPUT',
    VALIDATION_FAILED = 'VALIDATION_FAILED',
    MALFORMED_PAYLOAD = 'MALFORMED_PAYLOAD',
    UNSUPPORTED_OPERATION = 'UNSUPPORTED_OPERATION',
    UNAUTHENTICATED = 'UNAUTHENTICATED',
    FORBIDDEN_ACTION = 'FORBIDDEN_ACTION',
    RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
    RESOURCE_ALREADY_EXISTS = 'RESOURCE_ALREADY_EXISTS',
    RESOURCE_CONFLICT = 'RESOURCE_CONFLICT',
    BUSINESS_RULE_VIOLATION = 'BUSINESS_RULE_VIOLATION',
    OPERATION_NOT_ALLOWED = 'OPERATION_NOT_ALLOWED',
    STATE_MISMATCH = 'STATE_MISMATCH',
    SERVICE_UNAVAILABLE = 'SERVICE_UNAVAILABLE',
    DEPENDENCY_FAILURE = 'DEPENDENCY_FAILURE',
    TIMEOUT = 'TIMEOUT',
    INTERNAL_ERROR = 'INTERNAL_ERROR',
    DATA_ACCESS_ERROR = 'DATA_ACCESS_ERROR',
    SERIALIZATION_ERROR = 'SERIALIZATION_ERROR',
    UNKNOWN_ERROR = 'UNKNOWN_ERROR',
}